requirements.txt
streamlit
azure-ai-documentintelligence
python-dotenv
openai
Pillow
pdfminer
pymupdf